Split `BezierCurveTests` to multiple files
